requirements.txt
appnope==0.1.0
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
MarkupSafe==1.0
parso==0.1.0
passlib==1.7.1
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy-1.3.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
gunicorn==19.7.1
pandas==0.24.2
Flask==1.0.2
Flask_SQLAlchemy==2.3.2

